village-radio/scripts/requirements.txt
spotipy>=2.24.0
tidalapi>=0.7.0
pyyaml>=6.0
ruamel.yaml>=0.18.0
